Update test expectations and examples for weak typing
Adjust six existing test files for the new float literal behavior:
- test_codegen: update expected type names and error messages
- test_vec_constructors: float literals now infer as float32 (not error)
- test_mat_constructors: float literals now infer as float32 (not error)
- test_generics: update expected types for generic function resolution
- test_vec: adjust vector construction expectations
- test_mat_lite: adjust matrix fill expectations

Simplify tile examples now that literals adapt precision:
- example_tile_convolution: remove explicit wp.float64() wrapping
- example_tile_filtering: remove explicit wp.float64() wrapping

+++ warp/examples/tile/example_tile_convolution.py
@@ -31,7 +31,7 @@ BLOCK_DIM = 64
 TILE_M = 1
 TILE_N = 128
 
-scale = wp.vec2d(wp.float64(1 / TILE_N), wp.float64(1 / TILE_N))
+scale = wp.vec2d(1 / TILE_N, 1 / TILE_N)
 
 
 @wp.func
