+++ warp/tests/matrix/test_mat_lite.py
@@ -27,10 +27,10 @@ def test_matrix_constructor_value_func():
     b = wp.matrix(a, shape=(2, 2))
     c = mat32d()
     d = mat32d(c, shape=(3, 2))
-    e = mat32d(wp.float64(1.0), wp.float64(2.0), wp.float64(1.0), wp.float64(2.0), wp.float64(1.0), wp.float64(2.0))
+    e = mat32d(1.0, 2.0, 1.0, 2.0, 1.0, 2.0)
     f = wp.matrix_from_rows(
-        wp.vec3d(wp.float64(1.0), wp.float64(2.0), wp.float64(3.0)),
-        wp.vec3d(wp.float64(1.0), wp.float64(2.0), wp.float64(3.0)),
+        wp.vec3d(1.0, 2.0, 3.0),
+        wp.vec3d(1.0, 2.0, 3.0),
     )
     g = wp.types.matrix(1.0, shape=(3, 2))
